test: correct relationship between filename and test codes
